+++ src/index.ts
@@ -26,6 +26,7 @@ import BaseConnectionHealthPolicy from './connectionhealthpolicy/BaseConnectionH
 import BaseTask from './task/BaseTask';
 import BitrateParameters from './videouplinkbandwidthpolicy/BitrateParameters';
 import BrowserBehavior from './browserbehavior/BrowserBehavior';
+import CanvasVideoFrameBuffer from './videoframeprocessor/CanvasVideoFrameBuffer';
 import CheckAudioConnectivityFeedback from './meetingreadinesschecker/CheckAudioConnectivityFeedback';
 import CheckAudioInputFeedback from './meetingreadinesschecker/CheckAudioInputFeedback';
 import CheckAudioOutputFeedback from './meetingreadinesschecker/CheckAudioOutputFeedback';
@@ -80,11 +81,14 @@ import DefaultSimulcastUplinkPolicy from './videouplinkbandwidthpolicy/DefaultSi
 import DefaultStatsCollector from './statscollector/DefaultStatsCollector';
 import DefaultTransceiverController from './transceivercontroller/DefaultTransceiverController';
 import DefaultVideoCaptureAndEncodeParameter from './videocaptureandencodeparameter/DefaultVideoCaptureAndEncodeParameter';
+import DefaultVideoFrameProcessorPipeline from './videoframeprocessor/DefaultVideoFrameProcessorPipeline';
 import DefaultVideoStreamIdSet from './videostreamidset/DefaultVideoStreamIdSet';
 import DefaultVideoStreamIndex from './videostreamindex/DefaultVideoStreamIndex';
 import DefaultVideoTile from './videotile/DefaultVideoTile';
 import DefaultVideoTileController from './videotilecontroller/DefaultVideoTileController';
 import DefaultVideoTileFactory from './videotilefactory/DefaultVideoTileFactory';
+import DefaultVideoTransformDevice from './videoframeprocessor/DefaultVideoTransformDevice';
+import DefaultVideoTransformDeviceObserver from './videoframeprocessor/DefaultVideoTransformDeviceObserver';
 import DefaultVolumeIndicatorAdapter from './volumeindicatoradapter/DefaultVolumeIndicatorAdapter';
 import DefaultWebSocketAdapter from './websocketadapter/DefaultWebSocketAdapter';
 import Device from './devicecontroller/Device';
@@ -146,6 +150,7 @@ import NoOpVideoElementFactory from './videoelementfactory/NoOpVideoElementFacto
 import NoVideoDownlinkBandwidthPolicy from './videodownlinkbandwidthpolicy/NoVideoDownlinkBandwidthPolicy';
 import NoVideoUplinkBandwidthPolicy from './videouplinkbandwidthpolicy/NoVideoUplinkBandwidthPolicy';
 import None from './maybe/None';
+import NoopVideoFrameProcessor from './videoframeprocessor/NoopVideoFrameProcessor';
 import NotFoundError from './devicecontroller/NotFoundError';
 import NotReadableError from './devicecontroller/NotReadableError';
 import OpenSignalingConnectionTask from './task/OpenSignalingConnectionTask';
@@ -210,6 +215,10 @@ import VideoAdaptiveProbePolicy from './videodownlinkbandwidthpolicy/VideoAdapti
 import VideoCaptureAndEncodeParameter from './videocaptureandencodeparameter/VideoCaptureAndEncodeParameter';
 import VideoDownlinkBandwidthPolicy from './videodownlinkbandwidthpolicy/VideoDownlinkBandwidthPolicy';
 import VideoElementFactory from './videoelementfactory/VideoElementFactory';
+import VideoFrameBuffer from './videoframeprocessor/VideoFrameBuffer';
+import VideoFrameProcessor from './videoframeprocessor/VideoFrameProcessor';
+import VideoFrameProcessorPipeline from './videoframeprocessor/VideoFrameProcessorPipeline';
+import VideoFrameProcessorPipelineObserver from './videoframeprocessor/VideoFrameProcessorPipelineObserver';
 import VideoInputDevice from './devicecontroller/VideoInputDevice';
 import VideoLogEvent from './statscollector/VideoLogEvent';
 import VideoQualitySettings from './devicecontroller/VideoQualitySettings';
@@ -222,6 +231,7 @@ import VideoTileController from './videotilecontroller/VideoTileController';
 import VideoTileControllerFacade from './videotilecontroller/VideoTileControllerFacade';
 import VideoTileFactory from './videotilefactory/VideoTileFactory';
 import VideoTileState from './videotile/VideoTileState';
+import VideoTransformDevice from './devicecontroller/VideoTransformDevice';
 import VideoUplinkBandwidthPolicy from './videouplinkbandwidthpolicy/VideoUplinkBandwidthPolicy';
 import VoiceFocusConfig from './voicefocus/VoiceFocusConfig';
 import VoiceFocusDeviceOptions from './voicefocus/VoiceFocusDeviceOptions';
@@ -265,6 +275,7 @@ export {
   BaseTask,
   BitrateParameters,
   BrowserBehavior,
+  CanvasVideoFrameBuffer,
   CheckAudioConnectivityFeedback,
   CheckAudioInputFeedback,
   CheckAudioOutputFeedback,
@@ -319,11 +330,14 @@ export {
   DefaultStatsCollector,
   DefaultTransceiverController,
   DefaultVideoCaptureAndEncodeParameter,
+  DefaultVideoFrameProcessorPipeline,
   DefaultVideoStreamIdSet,
   DefaultVideoStreamIndex,
   DefaultVideoTile,
   DefaultVideoTileController,
   DefaultVideoTileFactory,
+  DefaultVideoTransformDevice,
+  DefaultVideoTransformDeviceObserver,
   DefaultVolumeIndicatorAdapter,
   DefaultWebSocketAdapter,
   Device,
@@ -385,6 +399,7 @@ export {
   NoVideoDownlinkBandwidthPolicy,
   NoVideoUplinkBandwidthPolicy,
   None,
+  NoopVideoFrameProcessor,
   NotFoundError,
   NotReadableError,
   OpenSignalingConnectionTask,
@@ -449,6 +464,10 @@ export {
   VideoCaptureAndEncodeParameter,
   VideoDownlinkBandwidthPolicy,
   VideoElementFactory,
+  VideoFrameBuffer,
+  VideoFrameProcessor,
+  VideoFrameProcessorPipeline,
+  VideoFrameProcessorPipelineObserver,
   VideoInputDevice,
   VideoLogEvent,
   VideoQualitySettings,
@@ -461,6 +480,7 @@ export {
   VideoTileControllerFacade,
   VideoTileFactory,
   VideoTileState,
+  VideoTransformDevice,
   VideoUplinkBandwidthPolicy,
   VoiceFocusConfig,
   VoiceFocusDeviceOptions,